complete mnemonic phrase encrypt and reuse
