+++ crates/rooch-rpc-server/src/lib.rs
@@ -185,7 +185,8 @@ pub async fn run_start_server(opt: &RoochOpt, mut server_opt: ServerOpt) -> Resu
             let result = generate_new_key_pair(
                 None, None, None, // TODO: provide a password option for server start
             )?;
-            let kp: RoochKeyPair = retrieve_key_pair(&result.result.encryption, None)?; // TODO: provide a password option for server start
+            let kp: RoochKeyPair =
+                retrieve_key_pair(&result.key_pair_data.private_key_encryption, None)?; // TODO: provide a password option for server start
             server_opt.sequencer_keypair = Some(kp.copy());
             server_opt.proposer_keypair = Some(kp.copy());
             server_opt.relayer_keypair = Some(kp.copy());
