+++ crates/rooch/src/commands/account/commands/update.rs
@@ -3,10 +3,7 @@
 
 use clap::Parser;
 use move_core_types::account_address::AccountAddress;
-use rooch_key::{
-    key_derive::{retrieve_key_pair, verify_password},
-    keystore::AccountKeystore,
-};
+use rooch_key::key_derive::{retrieve_key_pair, verify_password};
 use rooch_rpc_api::jsonrpc_types::ExecuteTransactionResponseView;
 use rooch_types::{
     address::RoochAddress,
@@ -16,6 +13,7 @@ use rooch_types::{
 use rpassword::prompt_password;
 
 use crate::cli_types::{CommandAction, WalletContextOptions};
+use rooch_key::keystore::account_keystore::AccountKeystore;
 use std::str::FromStr;
 
 /// Update an address with a new keypair from coin id to rooch.keystore
@@ -48,10 +46,9 @@ impl CommandAction<ExecuteTransactionResponseView> for UpdateCommand {
                 None,
             )
         } else {
-            let password = prompt_password(
-                "Enter the password saved in client config to update address with a new encryption data:",
-            )
-            .unwrap_or_default();
+            let password =
+                prompt_password("Enter the password to update address with a new encryption data:")
+                    .unwrap_or_default();
             let is_verified = verify_password(
                 Some(password.clone()),
                 context
