+++ crates/rooch/src/commands/session_key/commands/create.rs
@@ -4,7 +4,8 @@
 use crate::cli_types::{TransactionOptions, WalletContextOptions};
 use clap::Parser;
 use moveos_types::module_binding::MoveFunctionCaller;
-use rooch_key::{key_derive::verify_password, keystore::AccountKeystore};
+use rooch_key::key_derive::verify_password;
+use rooch_key::keystore::account_keystore::AccountKeystore;
 use rooch_types::{
     address::RoochAddress,
     error::{RoochError, RoochResult},
@@ -49,10 +50,8 @@ impl CreateCommand {
         let session_auth_key = if context.client_config.is_password_empty {
             context.keystore.generate_session_key(&sender, None)?
         } else {
-            let password = prompt_password(
-                "Enter the password saved in client config to create a new key pair:",
-            )
-            .unwrap_or_default();
+            let password =
+                prompt_password("Enter the password to create a new key pair:").unwrap_or_default();
             let is_verified = verify_password(
                 Some(password.clone()),
                 context
@@ -87,10 +86,8 @@ impl CreateCommand {
         let result = if context.client_config.is_password_empty {
             context.sign_and_execute(sender, action, None).await?
         } else {
-            let password = prompt_password(
-                "Enter the password saved in client config to create a new key pair:",
-            )
-            .unwrap_or_default();
+            let password =
+                prompt_password("Enter the password to create a new key pair:").unwrap_or_default();
             let is_verified = verify_password(
                 Some(password.clone()),
                 context
